src/layers/core/python/requirements.txt
pytz~=2021.1
aws-lambda-powertools==1.22.0
peewee==3.14.8
PyJWT~=2.4.0
pysftp==0.2.9
requests
aws-xray-sdk
openpyxl
croniter